Move T1/T2 tab logic into T1T2TabController
